boot/openfirmware: fix truncated OF handle, enable the FPU on sparc
Two bugs that together looked like a broken firmware. With both fixed the
loader reaches its interactive boot menu on the QEMU sun4u machine.

devices.cpp stored an Open Firmware instance handle in an int:

	int handle = of_open(devicePath);

of_open returns intptr_t. On sparc64 the assignment drops the high half,
and the value then sign-extends back into something that is not a valid
handle. The firmware dereferenced it and took a data access exception
inside itself, so the trap PC pointed into OpenBIOS and the failure
presented convincingly as a broken seek or read. On 32-bit PowerPC int
and intptr_t coincide, which is why this was invisible there.

With that fixed the trap moved into our own code, at a load into %f0:
trap 0x20, fp_disabled. SPARC V9 gates floating point behind PSTATE.PEF
(bit 4) and FPRS.FEF (bit 2), and an FP instruction with either bit clear
traps; see the UltraSPARC-IIi User's Manual section A.4. Open Firmware
does not set them for a client program, and the loader does use floating
point -- the menu formats partition sizes with %f, and GCC emits FP
register loads to move small structures, which is what faulted first.
_start now enables both before the constructors run. The VIS graphics
instructions share the FP register file and the same two enables.

The loader now enumerates every device, scans partitions, probes BFS,
FAT32 and TAR, and draws its menu. It reports no supported boot device,
which is correct: the test image carries ext2. A BFS volume with a kernel
is the next step.

Two remaining QEMU-only quirks, both about empty removable drives and both
avoided by attaching media: opening an empty CD-ROM traps 0x34 inside
OpenBIOS, and scanning an empty floppy hangs.

+++ src/system/boot/platform/openfirmware/devices.cpp
@@ -119,7 +119,13 @@ platform_add_boot_device(struct stage2_args *args, NodeList *devicesList)
 			file[0] = '\0';
 	}
 
-	int handle = of_open(devicePath);
+	// An Open Firmware handle is intptr_t, not int. Truncating it to 32 bits
+	// happened to be harmless on 32-bit PowerPC, but on sparc64 the high bits
+	// are lost and the value then sign-extends back into something that is not
+	// a valid instance handle. The firmware dereferences it and takes a data
+	// access exception inside itself, which looks for all the world like a
+	// broken seek or read.
+	intptr_t handle = of_open(devicePath);
 	if (handle == OF_FAILED) {
 		puts("\t\t(open failed)");
 		return B_ERROR;
